Fixing format issue

+++ lib/infrastructure/income_statement/models/income_statement_model.dart
@@ -31,7 +31,7 @@ class IncomeStatementModel {
   final int? generalAndAdministrativeExpenses;
   final int? sellingAndMarketingExpenses;
   final int? sellingGeneralAndAdministrativeExpenses;
-  final int? otherExpenses;
+  final double? otherExpenses;
   final int? operatingExpenses;
   final int? costAndExpenses;
   final int? interestIncome;
